[story/ECP-624] - renames FWY => DQT
