+++ .env.development
@@ -5,7 +5,5 @@ DFE_SIGN_IN_IDENTIFIER=teacherpayments
 DFE_SIGN_IN_API_CLIENT_ID=teacherpayments
 DFE_SIGN_IN_API_ENDPOINT=https://pp-api.signin.education.gov.uk
 
-DQT_CLIENT_HOST=https://s118d01funcappfapp.azurewebsites.net/
-
 ADMIN_ALLOWED_IPS=::1,127.0.0.1
 ENVIRONMENT_NAME=local
