+++ .env.production
@@ -5,8 +5,6 @@ DFE_SIGN_IN_IDENTIFIER=teacherpayments
 DFE_SIGN_IN_API_CLIENT_ID=teacherpayments
 DFE_SIGN_IN_API_ENDPOINT=https://pp-api.signin.education.gov.uk
 
-DQT_CLIENT_HOST=https://s118d01funcappfapp.azurewebsites.net/
-
 ENVIRONMENT_NAME=review
 
 RAILS_SERVE_STATIC_FILES=true
