+++ .env.test
@@ -2,18 +2,13 @@ DFE_SIGN_IN_API_CLIENT_ID=teacherpayments
 DFE_SIGN_IN_API_SECRET=secret
 DFE_SIGN_IN_API_ENDPOINT=https://example.com
 
-DQT_CLIENT_HEADERS={"header": "value"}
-DQT_CLIENT_HOST=http://dqt.com
-DQT_CLIENT_PARAMS={"param": "value"}
-DQT_CLIENT_PORT=13
-
-FWY_BEARER_BASE_URL=https://login.microsoftonline.com/blah/oauth2/v2.0/token
-FWY_BEARER_GRANT_TYPE=client_credentials
-FWY_BEARER_SCOPE=https://test.dynamics.com/.default
-FWY_BEARER_CLIENT_ID=1234
-FWY_BEARER_CLIENT_SECRET=5678
-FWY_BASE_URL=https://test-api-customerengagement.platform.education.gov.uk/dqt-crm/v1/
-FWY_SUBSCRIPTION_KEY=09876432
+DQT_BEARER_BASE_URL=https://login.microsoftonline.com/blah/oauth2/v2.0/token
+DQT_BEARER_GRANT_TYPE=client_credentials
+DQT_BEARER_SCOPE=https://test.dynamics.com/.default
+DQT_BEARER_CLIENT_ID=1234
+DQT_BEARER_CLIENT_SECRET=5678
+DQT_BASE_URL=https://test-api-customerengagement.platform.education.gov.uk/dqt-crm/v1/
+DQT_SUBSCRIPTION_KEY=09876432
 
 ADMIN_ALLOWED_IPS=::1,127.0.0.1
