+++ lib/dqt/client.rb
@@ -1,4 +1,4 @@
-module Fwy
+module Dqt
   class Client
     attr_reader :adapter
 
@@ -13,7 +13,7 @@ module Fwy
     end
 
     def connection
-      @connection ||= Faraday.new(ENV["FWY_BASE_URL"]) do |c|
+      @connection ||= Faraday.new(ENV["DQT_BASE_URL"]) do |c|
         c.request :authorization, "Bearer", Bearer.get_auth_token
         c.request :json
         c.response :json, content_type: "application/json"
