+++ lib/dqt/error.rb
@@ -1,4 +1,4 @@
-module Fwy
+module Dqt
   class Error < StandardError
   end
 end
