+++ lib/dqt/resource.rb
@@ -1,4 +1,4 @@
-module Fwy
+module Dqt
   class Resource
     attr_reader :client
 
@@ -36,7 +36,7 @@ module Fwy
     private
 
     def default_headers
-      {"Ocp-Apim-Subscription-Key" => ENV["FWY_SUBSCRIPTION_KEY"]}
+      {"Ocp-Apim-Subscription-Key" => ENV["DQT_SUBSCRIPTION_KEY"]}
     end
   end
 end
