--- lib/fwy.rb
@@ -1,7 +1,0 @@
-require_relative "fwy/error"
-require_relative "fwy/bearer"
-require_relative "fwy/client"
-require_relative "fwy/object"
-require_relative "fwy/objects/teacher"
-require_relative "fwy/resource"
-require_relative "fwy/resources/teacher"
