+++ spec/rails_helper.rb
@@ -67,8 +67,7 @@ RSpec.configure do |config|
   config.include ActiveJob::TestHelper
   config.include DfeSignInHelpers
   config.include GeckoboardHelpers
-  # config.include DqtHelpers
-  config.include FwyHelpers
+  config.include DqtHelpers
   config.include OrdnanceSurveyHelpers
   config.include FixtureHelpers
   config.include IsExpectedWithBlock
